third_party/blink/web_tests/wpt_internal/css/css-masking/parsing/webkit-mask-position-x-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Non-standard CSS Masking: parsing background-position-x with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-backgrounds-4/#propdef-background-position-x">
<meta name="assert" content="background-position-x supports only the grammar '[ center | [ left | right | x-start | x-end ]? <length-percentage>? ]#'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("-webkit-mask-position-x", "top");
test_invalid_value("-webkit-mask-position-x", "bottom");
test_invalid_value("-webkit-mask-position-x", "y-start");
test_invalid_value("-webkit-mask-position-x", "y-end");
test_invalid_value("-webkit-mask-position-x", "center 10px");
test_invalid_value("-webkit-mask-position-x", "20% left");
test_invalid_value("-webkit-mask-position-x", "right left");
test_invalid_value("-webkit-mask-position-x", "x-start center");
test_invalid_value("-webkit-mask-position-x", "left, center right");
</script>
</body>
</html>
